refactor(serve): close remaining boundary gaps + move helpers to dedicated module
Review feedback for #524:

1. Architecture/Code Quality MEDIUM — helpers don't belong in mod.rs
   `mod.rs` is the HTTP entrypoint (router setup, server boot). The
   katulong protocol helpers (extract_session_id and create_or_find_session)
   were sitting alongside it. Moved both into a new
   `serve/katulong_proxy.rs` submodule. Callers now do
   `super::katulong_proxy::create_or_find_session(...)`.
   `extract_session_id` was downgraded from `pub(super)` to private
   `fn` since only `create_or_find_session` (in the same file) calls
   it now (per Correctness LOW).

2. Architecture/Vision MEDIUM — boundary not fully closed
   Three remaining inline `format!` URL constructions were not migrated
   to sipag-core helpers in #521's main commit:
   - board.rs:537: `format!("/sessions/by-id/{}/status", sid)` in
     proxy_session_status
   - htmx.rs:815: claude-transcript URL with limit query param
   - htmx.rs:915: claude-respond POST URL
   Added two new URL builders to sipag-core (claude_transcript_url,
   claude_respond_url) and refactored proxy_get to take a URL +
   `&Host` directly so callers build URLs via the existing helpers.
   Both proxy callers (proxy_sessions, proxy_session_status) now
   resolve the host themselves and pass `sessions_url(...)` /
   `status_url(...)` through.

What changed
- sipag-core/src/katulong.rs: +2 URL builders (claude_transcript_url,
  claude_respond_url), +2 unit tests
- sipag/src/serve/katulong_proxy.rs: new module with the two
  protocol helpers (~115 lines, split out from mod.rs)
- sipag/src/serve/mod.rs: helpers removed; module declaration added;
  unused StatusCode import dropped
- sipag/src/serve/board.rs: proxy_get takes (url: &str, host: &Host)
  instead of (path: &str, host_id: &str). Callers do host lookup +
  URL construction themselves
- sipag/src/serve/htmx.rs: transcript and respond URLs now use the
  new builders

Out of scope (left as future follow-ups)
- DispatchResponse.session_id remains Option<String>. The comment
  there says "flatten on the next API revision"; this PR is internal
  cleanup, not an API revision, so deferred.
- The pre-existing security findings (HTTP error body forwarding,
  session_id format validation, unbounded JSON, gemma4 input
  pass-through) are tracked separately in subsequent issues.

Validation
  cargo build: clean
  make dev: 28 + 17 + 22 + 167 (was 165; +2 URL-builder tests) +
            7 tests pass; clippy clean; fmt clean
  cargo run --example probe_by_id: full e2e against local katulong ✓
